requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
weasyprint>=60.0



